+++ teamized/app/shadcn/components/ui/sonner.tsx
@@ -5,16 +5,13 @@ import {
     OctagonXIcon,
     TriangleAlertIcon,
 } from 'lucide-react';
-import { useTheme } from 'next-themes';
 import React from 'react';
 import { Toaster as Sonner, type ToasterProps } from 'sonner';
 
-const Toaster = ({ ...props }: ToasterProps) => {
-    const { theme = 'system' } = useTheme();
-
+const Toaster = ({ theme, ...props }: ToasterProps) => {
     return (
         <Sonner
-            theme={theme as ToasterProps['theme']}
+            theme={theme}
             className="tw:toaster tw:group"
             icons={{
                 success: <CircleCheckIcon className="tw:size-4" />,